Aggregate over calendar periods

It computes summary statistics for a tsibble over calendar periods, usually used in combination of group_by.

# S3 method for tbl_ts
summarise(.data, ...)

summarize(.data, ...)

Arguments

.data
A tsibble (of tbl_ts class).

...
Name-value pairs of summary functions. To aggregate tsibble over a certain calendar period, for example yearly aggregates, ~ year() needs passing to .... Please see details.

Value

A tsibble class when the ~ is present.

Details

It's S3 method implemented for tsibble() (tbl_ts) obtained from dplyr::summarise(). A formula with ~ followed by one of calendar component functions from base, lubridate and zoo specifies the period when summary functions are carried out. Currently ~ year() indicates yearly aggregates. ~ yearqtr() indicates quarterly aggregates. ~ yearmon() indicates monthly aggregates. ~ as_date() or as.Date() indicates daily aggregates.

See also

dplyr::summarise

Examples

   pkgs_ts <- as_tsibble(tidypkgs, key = key_vars(package), index = date)
   pkgs_ts %>%
     group_by(package) %>%
     summarise(avg_count = mean(count), month = ~ as.yearmon())

#> # A tsibble of 1MONTH time interval for package 
#> # A tibble: 144 x 3
#> # Groups:   package [?]
#>    package         month avg_count
#>  *   <chr> <S3: yearmon>     <dbl>
#>  1   dplyr      Jan 2015  1311.097
#>  2   dplyr      Feb 2015  1727.821
#>  3   dplyr      Mar 2015  1630.000
#>  4   dplyr      Apr 2015  1869.067
#>  5   dplyr      May 2015  1843.290
#>  6   dplyr      Jun 2015  2224.567
#>  7   dplyr      Jul 2015  1713.484
#>  8   dplyr      Aug 2015  1961.387
#>  9   dplyr      Sep 2015  2925.600
#> 10   dplyr      Oct 2015  2774.161
#> # ... with 134 more rows